requirements.txt
celery
django
django-money
django-filter
django-redis
djangorestframework
drf-spectacular
factory-boy
faker
psycopg2
pyxirr
